requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread==5.12.0
oauth2client==4.1.3
pillow==10.2.0
pyinstaller==6.3.0 